+++ ReleaseNotes.md
@@ -1,5 +1,8 @@
 # Release Notes for Qlik Sense Rest Client
 
+## v1.14.0
+* **NEW FEATURE:** Added overloads for `IRestClient.Put` and `IRestClient.PutAsync` that takes as argument an instance of `HttpContent`.
+
 ## v1.13.0
 * **NEW FEATURE:** Added overloads for `IRestClient.Post` and `IRestClient.PostAsync` that takes as argument an instance of `HttpContent`.
 * **EXAMPLE:** Added example `QcsImportExtension` that illustrates how to import an extension to QCS.